文章浏览量定时同步，更新浏览量

+++ framework-general/src/main/java/com/cg/entity/Articles.java
@@ -27,9 +27,9 @@ public class Articles {
     //文章预览图片
     private String img;
     //点赞数
-    private Long likeNumber;
+    private Integer likeNumber;
     //预览数
-    private Long previewNumber;
+    private Integer previewNumber;
     //评论数
     private Long commentNumber;
     //登陆才允许评论(0为否，1为是)
